requirements.txt
altair==4.2.2
matplotlib==3.7.0
mplcursors==0.5.2
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
plotly==5.9.0
scikit_learn==1.2.1
seaborn==0.12.2
streamlit==1.22.0
